requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
distlib==0.3.6
Django==3.2.17
django-cors-headers==3.13.0
django-decouple==2.1
django-filter==23.1
django-scheduler==0.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
filelock==3.10.7
icalendar==5.0.4
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
Markdown==3.4.3
mysqlclient==2.1.1
Pillow==9.4.0
pipenv==2023.3.20
platformdirs==3.2.0
prompt-toolkit==3.0.36
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
twilio==7.17.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
whitenoise==6.3.0
